# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bankcard OpenAPI definition
  version: v0.5
security:
  - Bearer Authentication: []
paths:
...
  /usercards/block:
    post:
      tags:
        - user-card-controller
      operationId: blockCardRequest
      requestBody:
        content:
          application/json:
            schema:
              type: integer
              format: int64
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/BlockRequestResponse"
...
  /cards/block:
    post:
      tags:
        - admin-cards-controller
      operationId: blockCard
      requestBody:
        content:
          application/json:
            schema:
              type: integer
              format: int64
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/CardResponse"
...
  /usercards/cards:
    get:
      tags:
        - user-card-controller
      operationId: myCards
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 5
        - name: sort
          in: query
          required: false
          schema:
            type: array
            default:
              - id
              - asc
            items:
              type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CardResponse"
  /usercards/balance:
    get:
      tags:
        - user-card-controller
      operationId: cardBalance
      parameters:
        - name: cardId
          in: query
          required: false
          schema:
            type: integer
            format: int64
            default: 0
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: number
...
components:
  schemas:
...
    BlockRequestResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cardId:
          type: integer
          format: int64
        maskedNumber:
          type: string
        status:
          type: string
          enum:
            - APPROVED
            - PENDING
...
    CardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maskedNumber:
          type: string
        expiryDate:
          type: string
          format: date
        status:
          type: string
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
        balance:
          type: number
...
  securitySchemes:
    Bearer Authentication:
      type: http
      scheme: Bearer
      bearerFormat: JWT